Split out ProcessReceiveCommand.

+++ usb64/usb64UnitTests/UnfDebugger.cs
@@ -121,7 +121,7 @@ namespace usb64UnitTests
         {
             var packetBody = "abcd";
             var command = new Unf.ReceiveCommandPacket();
-            command.Type = (int)Unf.ReceiveCommandPacket.CommandType.TEXT;
+            command.Type = Unf.ReceiveCommandPacket.CommandType.TEXT;
             command.Body = Encoding.ASCII.GetBytes(packetBody);
             var output = new Unf.Debugger().ProcessReceiveCommand(command.Encode());
             Assert.AreEqual(packetBody, output);
@@ -136,7 +136,7 @@ namespace usb64UnitTests
             random.NextBytes(packetBody);
             var command = new Unf.ReceiveCommandPacket
             {
-                Type = (int)Unf.ReceiveCommandPacket.CommandType.BINARY,
+                Type = Unf.ReceiveCommandPacket.CommandType.BINARY,
                 Body = packetBody
             };
             var output = new Unf.Debugger().ProcessReceiveCommand(command.Encode());
@@ -153,7 +153,7 @@ namespace usb64UnitTests
             var packetBody = new byte[] { 0x00, 0x01, 0x02, 0x03 };
             var command = new Unf.ReceiveCommandPacket
             {
-                Type = (int)Unf.ReceiveCommandPacket.CommandType.BINARY,
+                Type = Unf.ReceiveCommandPacket.CommandType.BINARY,
                 Body = packetBody
             };
             var commandWithPadding = new List<byte>();
@@ -182,7 +182,7 @@ namespace usb64UnitTests
 
             var command = new Unf.ReceiveCommandPacket
             {
-                Type = (int)Unf.ReceiveCommandPacket.CommandType.SCREENSHOT_HEADER,
+                Type = Unf.ReceiveCommandPacket.CommandType.SCREENSHOT_HEADER,
                 Body = packetBody.Encode()
             };
             var output = new Unf.Debugger().ProcessReceiveCommand(command.Encode());
